requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
Django==4.1
django-cors-headers==3.13.0
django-filter==22.1
django-phonenumber-field==6.3.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
phonenumbers==8.12.54
Pillow==9.2.0
platformdirs==2.5.2
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.12
